% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\docType{class}
\name{ExternalRule-class}
\alias{ExternalRule-class}
\title{External rule allow to use logical permutation vector that defines records
that vialate the business rule.}
\description{
External rule allow to use logical permutation vector that defines records
that vialate the business rule.
}
