% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\docType{methods}
\name{getValues,FieldRule,data.table-method}
\alias{getValues,FieldRule,data.table-method}
\title{Gets value(s) that contribute to rule vialation in a data.table}
\usage{
\S4method{getValues}{FieldRule,data.table}(rule, errors)
}
\arguments{
\item{rule}{data rule that will be used to determine how to make the vector of values}

\item{errors}{data.table that only contains records with errors}
}
\value{
vector of values that represent invalid records
}
\description{
Gets value(s) that contribute to rule vialation in a data.table
}
