% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{newConditionRule}
\alias{newConditionRule}
\title{Creates new condition rule}
\usage{
newConditionRule(field, condition)
}
\arguments{
\item{field}{name of the field to validate}

\item{condition}{sting containing parsable condition for validation}
}
\description{
Creates new condition rule
}
