% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{newEnumRule}
\alias{newEnumRule}
\title{Creates new enum rule}
\usage{
newEnumRule(field, enum)
}
\arguments{
\item{field}{name of the field to be validated}

\item{enum}{vector of allowed character values}
}
\value{
new enum rule
}
\description{
Creates new enum rule
}
