% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{newExternalRule}
\alias{newExternalRule}
\title{Creates new external rule, where calculation of the vector that defines erros
was done externally.}
\usage{
newExternalRule(field, errors)
}
\arguments{
\item{field}{name of the field that represents the error record}

\item{errors}{logical vector that defines records with errors}
}
\value{
new external rule
}
\description{
Creates new external rule, where calculation of the vector that defines erros
was done externally.
}
