% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-key.R
\name{newForeignKeyRule}
\alias{newForeignKeyRule}
\title{Factory for the ForeignKey rule}
\usage{
newForeignKeyRule(field, primary.key)
}
\arguments{
\item{field}{name of the foreign key field}

\item{primary.key}{vector of values from the referenced field}
}
\value{
rule for given parameters
}
\description{
Factory for the ForeignKey rule
}
