% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{newRegexRule}
\alias{newRegexRule}
\title{Creates new regex rule}
\usage{
newRegexRule(field, pattern)
}
\arguments{
\item{field}{name of the field to be validated}

\item{pattern}{regex pattern that field should comply with}
}
\value{
records that did not pass the check
}
\description{
Creates new regex rule
}
