% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required.R
\name{newRequiredRule}
\alias{newRequiredRule}
\title{Factory method for required rule}
\usage{
newRequiredRule(field)
}
\arguments{
\item{field}{name of the field to be validated}
}
\value{
new required rule
}
\description{
Factory method for required rule
}
