% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{newRulesContainer}
\alias{newRulesContainer}
\title{Creates new rules container from rules}
\usage{
newRulesContainer(source, ...)
}
\arguments{
\item{source}{data source of the data}

\item{...}{data rules that should be executed against the data}
}
\description{
Creates new rules container from rules
}
