% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{newSchema}
\alias{newSchema}
\title{Creates new rules container from rules}
\usage{
newSchema(source, schema, rules)
}
\arguments{
\item{source}{data source of the data}

\item{...}{data rules that should be executed against the data.
If first element is a list, it will be used as container of rules}
}
\value{
container with rules
}
\description{
Creates new rules container from rules
}
