% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{newUniqueRule}
\alias{newUniqueRule}
\title{Creates new unique rule}
\usage{
newUniqueRule(fields)
}
\arguments{
\item{fields}{vector of fields that define the unique constraint}
}
\value{
records that did not pass the check
}
\description{
Creates new unique rule
}
