% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\docType{class}
\name{PrimaryKeyRule-class}
\alias{PrimaryKeyRule-class}
\title{Class to validate the primariy key constraint
Finds and logs missing records in the referenced table}
\description{
Class to validate the primariy key constraint
Finds and logs missing records in the referenced table
}
