% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\docType{class}
\name{RecordRule-class}
\alias{RecordRule-class}
\title{Higher level class for errors that are at the record level, e.g. missing or duplicate}
\description{
Higher level class for errors that are at the record level, e.g. missing or duplicate
}
