% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{Schema}
\alias{Schema}
\title{Creates new rules container from rules}
\usage{
Schema(source, schema, rules = list())
}
\arguments{
\item{source}{data source of the data}

\item{schema}{list of data fields with names, types and rules (unique, regex, required)}

\item{rules}{list of data rules
If first element is a list, it will be used as container of rules}
}
\value{
container with rules
}
\description{
Creates new rules container from rules
}
