% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\docType{methods}
\name{validate,ConditionRule,data.table-method}
\alias{validate,ConditionRule,data.table-method}
\title{Validates data.table for regex field rule}
\usage{
\S4method{validate}{ConditionRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{Condition rule to validate data}

\item{dt}{Data for validation}
}
\value{
Records that did not pass the condition
}
\description{
Validates data.table for regex field rule
}
