% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\docType{methods}
\name{validate,DataRule,data.table-method}
\alias{validate,DataRule,data.table-method}
\title{Checks that rule and data.table are valid}
\usage{
\S4method{validate}{DataRule,data.table}(rule, dt)
}
\description{
callNextMethod() should be called from all overloads of this method
}
