% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\docType{methods}
\name{validate,EnumRule,data.table-method}
\alias{validate,EnumRule,data.table-method}
\title{Validates data.table for enum field rule}
\usage{
\S4method{validate}{EnumRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{enum rule}

\item{dt}{data for validation}
}
\value{
records that did not pass the check
}
\description{
Validates data.table for enum field rule
}
