% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\docType{methods}
\name{validate,ExternalRule,data.table-method}
\alias{validate,ExternalRule,data.table-method}
\title{Validates data.table for external rule}
\usage{
\S4method{validate}{ExternalRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{external rule}

\item{dt}{data for validation}
}
\value{
records that did not pass the check
}
\description{
Validates data.table for external rule
}
