% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-key.R
\docType{methods}
\name{validate,ForeignKeyRule,data.table-method}
\alias{validate,ForeignKeyRule,data.table-method}
\title{Validates data.table for foreign key data rule}
\usage{
\S4method{validate}{ForeignKeyRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{foreign key rule}

\item{dt}{data for validation}
}
\value{
records that did not pass the check
}
\description{
Validates data.table for foreign key data rule
}
