% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\docType{methods}
\name{validate,RegexRule,data.table-method}
\alias{validate,RegexRule,data.table-method}
\title{Validates data.table for regex field rule}
\usage{
\S4method{validate}{RegexRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{regex rule}

\item{dt}{data for validation}
}
\value{
records that did not pass the check
}
\description{
Validates data.table for regex field rule
}
