% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required.R
\docType{methods}
\name{validate,RequiredRule,data.table-method}
\alias{validate,RequiredRule,data.table-method}
\title{Validates data.table for the empty field rule}
\usage{
\S4method{validate}{RequiredRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{required rule}

\item{dt}{data for validation}
}
\value{
records that did not pass the check
}
\description{
Validates data.table for the empty field rule
}
