% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{methods}
\name{validate,rulesContainer,data.table-method}
\alias{validate,rulesContainer,data.table-method}
\title{Validates data.table against the container of rules}
\usage{
\S4method{validate}{rulesContainer,data.table}(rule, dt, conn = NULL,
  url.pattern = NA_character_)
}
\arguments{
\item{rule}{container with data rules}

\item{dt}{data.table that should be validated}

\item{conn}{connection to the database}

\item{url.pattern}{if provided, it will be used to generated url}
}
\value{
list of boolen values for each successful write to the database
}
\description{
Validates data.table against the container of rules
}
