% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{validate}
\alias{validate}
\title{Validates data in a data.table}
\usage{
validate(rule, dt)
}
\arguments{
\item{rule}{data rule that will be used to find records with errors}

\item{dt}{data to be validated}
}
\value{
subset of the original data that contains errors
}
\description{
Validates data in a data.table
}
