% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{validateRules}
\alias{validateRules}
\title{Validates data using rules within the rule container}
\usage{
validateRules(conn, container, dt, url.pattern = NA_character_)
}
\arguments{
\item{conn}{connection to the database}

\item{container}{container with rules}

\item{dt}{data to be validated}

\item{url.pattern}{if provided, it will be used to generated url}
}
\description{
Validates data using rules within the rule container
}
