% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-quality.R
\name{logWrongValues}
\alias{logWrongValues}
\title{Functions logs data.table with incrorrect records to the database}
\usage{
logWrongValues(conn, source, type, rule, refs, values, urls = rep(NA,
  length(values)))
}
\arguments{
\item{conn}{DBI connection}

\item{source}{name of the source system that contains incorrect record}

\item{type}{type of the error should be one of those: Orphan, Wrong, Missing, Duplicate}

\item{rule}{business rule that record violated}

\item{refs}{correlation references of records, e.g. primary key values}

\item{values}{wrong values}

\item{urls}{vector of urls to an object in the system, if not provided defaulted to NAs}
}
\value{
TRUE if records are saved successfuly
}
\description{
Functions logs data.table with incrorrect records to the database
}
