% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\docType{methods}
\name{validate,DataRule,data.table-method}
\alias{validate,DataRule,data.table-method}
\title{Checks that rule and data.table are valid
callNextMethod() should be called from all overloads of this method}
\usage{
\S4method{validate}{DataRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{data rule that will be used to find records with errors}

\item{dt}{data to be validated}
}
\value{
subset of the original data that contains errors
}
\description{
Checks that rule and data.table are valid
callNextMethod() should be called from all overloads of this method
}
