% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R, R/condition.R, R/container.R,
%   R/enum.R, R/external.R, R/foreign-key.R, R/regex.R, R/required.R,
%   R/schema.R, R/unique.R
\docType{methods}
\name{validate}
\alias{validate}
\alias{validate,DataRule,data.table-method}
\alias{validate,ConditionRule,data.table-method}
\alias{validate,RulesContainer,data.table-method}
\alias{validate,EnumRule,data.table-method}
\alias{validate,ExternalRule,data.table-method}
\alias{validate,ForeignKeyRule,data.table-method}
\alias{validate,RegexRule,data.table-method}
\alias{validate,RequiredRule,data.table-method}
\alias{validate,Schema,data.table-method}
\alias{validate,UniqueRule,data.table-method}
\title{Validates data in a data.table}
\usage{
validate(rule, dt, ...)

\S4method{validate}{DataRule,data.table}(rule, dt)

\S4method{validate}{ConditionRule,data.table}(rule, dt)

\S4method{validate}{RulesContainer,data.table}(rule, dt, conn = NULL,
  url.pattern = NA_character_)

\S4method{validate}{EnumRule,data.table}(rule, dt)

\S4method{validate}{ExternalRule,data.table}(rule, dt)

\S4method{validate}{ForeignKeyRule,data.table}(rule, dt)

\S4method{validate}{RegexRule,data.table}(rule, dt)

\S4method{validate}{RequiredRule,data.table}(rule, dt)

\S4method{validate}{Schema,data.table}(rule, dt, conn = NULL,
  url.pattern = NA_character_)

\S4method{validate}{UniqueRule,data.table}(rule, dt)
}
\arguments{
\item{rule}{data rule that will be used to find records with errors}

\item{dt}{data to be validated}

\item{...}{allows to extend validate function for new data rules}

\item{conn}{connection to the database}

\item{url.pattern}{if provided, it will be used to generated url}
}
\value{
subset of the original data that contains errors

for `ConditionRule` records where values in target field 
    don't match specified condition.

for `RulesContainer` error records with value of key column, target column and rule type

for `EnumRule` records where values don't match given list of valid values.

for `ExternalRule` records defined by TRUE/FALSE vector of permutations.

for `ForeignKeyRule` orphan records based on the foreign key constraint.

for `RegexRule` records where values don't match given regular expression.

for `RequiredRule` records where target field is NA or zero-length.

for `Schema` same as RulesContainer

for `UniqueRule` records where values in target field(s) are duplicated.
}
\description{
Validates data in a data.table
}
