% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{methods}
\name{validateRules,DBIConnection,rulesContainer,data.table-method}
\alias{validateRules,DBIConnection,rulesContainer,data.table-method}
\title{Validates data.table against the container of rules}
\usage{
\S4method{validateRules}{DBIConnection,rulesContainer,data.table}(conn,
  container, dt)
}
\arguments{
\item{conn}{connection to the database}

\item{container}{container with data rules}

\item{dt}{data.table that should be validated}
}
\description{
Validates data.table against the container of rules
}
