% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{methods}
\name{validateRules,DBIConnection,rulesContainer,data.table-method}
\alias{validateRules,DBIConnection,rulesContainer,data.table-method}
\title{Validates data.table against the container of rules}
\usage{
\S4method{validateRules}{DBIConnection,rulesContainer,data.table}(conn,
  container, dt, url.pattern = NA_character_)
}
\arguments{
\item{conn}{connection to the database}

\item{container}{container with data rules}

\item{dt}{data.table that should be validated}

\item{url.pattern}{if provided, it will be used to generated url}
}
\value{
list of boolen values for each successful write to the database
}
\description{
Validates data.table against the container of rules
}
